Rework debug integration. Move TimeSystem into Core/.
